+++ migrations/20230413123627_balance_table.sql
@@ -3,8 +3,8 @@
 select 'up SQL query';
 create table if not exists public.balance (
     id serial primary key,
-    balance decimal,
-    withdrawal decimal,
+    balance float,
+    withdrawal float,
     user_id int unique,
     constraint FK_balance_user foreign key (user_id) references public.users (id)
 );
